Use the `EndToken` enum for `parse_sequence` (#721)
This is in preparation for merging `Token::Begin` and `Token::End` with
`Token::EnvName` because it was a bit weird to use `Token::Begin` and
`Token::End` with an associated `Env` as the stopping token for
`parse_sequence` when `parse_sequence` doesn't actually look at the
`Env`.

Also, this change makes error reporting slightly less awkward.